Implement BaseModuleStorage for storage initialization in preparation of rights calculation extension
