+++ src/main/java/dk/kb/license/storage/RightsModuleStorage.java
@@ -1,0 +1,9 @@
+package dk.kb.license.storage;
+
+/**
+ * Class for interacting with the part of DS-License which handles the calculation of rights, such as holdbacks and restrictions.
+ */
+public class RightsModuleStorage extends BaseModuleStorage{
+
+    //TODO: Implement methods for the rights module
+}
